==> code/statistical_analysis/03_analyzing_relative_body_mass.Rmd <==
---
title: "Analyzing body mass of recaptures"
output: html_document
knit: (function(inputFile, encoding) {
      out_dir <- "knitted_markdown_files";
      rmarkdown::render(inputFile,
                        encoding=encoding,
                        output_dir=file.path(dirname(inputFile), out_dir))})
---

**Goal**

Analyze how experimental *Heliconia* removal influences:

- The body mass of hummingbirds captured during both experimental periods (pre and post; 'recaptures')

Predict that, if birds stayed in the treatment focal areas despite our major calorie depletion, they might lose body mass.

**Problem**

Problem with mass as a response variable: how do I account for differences in structural size within and between species? There are lots of different hummingbird species captured in this study system! Larger birds will have larger changes in body mass.

Potential options considered and ultimately not chosen:

- Random effect of species (for analysis of all species)

- Fixed effect of species (for analysis of *Heliconia* specialists, since there are only two species; not enough levels for a random effect)

- Covariate for structural size. Wing length? Wing length squared? Wing length cubed? Adding covariate for structural size is a bit tricky, because I don't know what the allometric exponent should be.

**Approach**

Analyze hummingbird body mass *relative to* mass predicted for a given wing length (within a species).

1. Using data from all years of hummingbird captures (2018-2018), create allometric equations for each species: log(mass) ~ log(wing length)

2. For each recaptured hummingbird from the experiment (2016-2018), use the allometric equation to calculate predicted body mass for a given wing length. For each bird, I therefore need data about wing length and body mass.

3. Calculate relative mass as measured body mass divided by predicted body mass. Birds with values <1 are lighter than expected for their structural size; similarly, birds with values >1 are heavier than expected for their structural.

4. Analyze how *Heliconia* removal influences change in relative body mass.

```{r setup, include=FALSE}
knitr::opts_chunk$set(message = FALSE, warning = FALSE)

library(dplyr)
library(tidyr)
library(lubridate)
library(ggplot2)
library(purrr)
library(glmmTMB)
library(emmeans)
```

## Step 1: Import data and functions

**Import data**
```{r}
#Hummingbirds codes --> hummingbird species names
bird_species_names <- read.csv("../../data/import/data/Hummingbird_codes_species_list.csv")

#Get all capture data (2010-2018)
#Add color_id column because sometimes birds were not banded. Using colors instead in those cases. This ID will allow me to identify recaps between pre/post a given patch/year
captures_all_years <- read.csv("../../data/import/data/captures/Hummingbird_captures_2010-2018.csv") %>%
  left_join(bird_species_names) %>%
  rename(patch = site) %>%
  filter(status != "DR") %>% #Remove day recaptures
  mutate(bird_id = ifelse(is.na(band_number), paste(year, patch, bird_species, colors, sep = "_"), paste(year, patch, bird_species, band_number, sep = "_"))) %>% 
  mutate(year_patch = paste(year, patch, sep = "_")) %>%
  select(bird_id, everything())

#Resource data. Need to add quantitative treatment measure here (*Heliconia* calories removed/ha)
heto_removed <- read.csv("../../data/export/for_analysis/Heliconia_resources_removed_as_treatment.csv") %>%
  select(-area) %>%
  mutate(patch = as.character(patch))

#Hummingbirds captured pre AND post during Heliconia removal experiment (2016-2018)
recap_list <- read.csv("../../data/export/intermediate/Birds_recaptured_post.csv")

#Subset capture data to hummingbirds recaptured during Heliconia removal experiment (2016-2018)
#Add info about quantitative treatment (calories removed/ha)
recaptures_experiment <- captures_all_years %>%
  filter(bird_id %in% recap_list$bird_id) %>%
  mutate(exp_phase = factor(exp_phase, levels = c("capture_1", "capture_2"), labels = c("pre", "post"))) %>%
  mutate(control_treatment = factor(control_treatment, levels = c("control", "treatment")))
```

```{r, eval = FALSE, include = FALSE}
#Calculate how long after sunrise (5:30) bird was captured...predict that birds captured later in day have had more time to fill up on nectar
#ULTIMATELY DECIDED THAT THIS WAS GOING TO INTRODUCE UNNEEDED COMPLEXITY INTO MODEL
recaptures_experiment_sunrise <- recaptures_experiment %>%
  mutate(time = lubridate::hm(time)) %>%
  mutate(sunrise = lubridate::hm("5:30")) %>%
  mutate(hours_since_sunrise = as.numeric(time - sunrise)/60/60)
```

**Import functions**
```{r}
source("../../code/helper_functions/Modeling_helper_functions.R")
source("../../code/helper_functions/Plotting_helper_functions.R")
source("../../code/helper_functions/Extract_data_from_safely_and_quietly_lists.R")
source("../../code/helper_functions/Calculate_basic_summary_stats.R")
```

## Step 2: Make sure recaptured birds have morphometric measuremets

To calculate a relative body mass, I need wing length - the measure of structural size - and the measured body mass.

```{r}
#Identify recaptures that have wing measurements, average across captures if a recaptured bird has multiple wing measurements.
#If a recaptured bird is missing wing length measurement during one period, just use the other capture. This will let me include more data in subsequent analyses.
average_wing_lengths <- recaptures_experiment %>%
  select(bird_id, exp_phase, wing_length) %>%
  group_by(bird_id) %>%
  summarise(wing_length = mean(wing_length, na.rm = TRUE)) %>%
  ungroup()

#Remove wing lengths, replace with average lengths
#Remove any birds without any wing measurements; e.g., some birds didn't have wing length measurements pre or post (p29, bc we got very busy at banding table)
#Remove any mass measurements where bird fed sugar water before mass was measured
recaptures_experiment_measurements <- recaptures_experiment %>%
  select(-wing_length) %>%
  left_join(average_wing_lengths) %>%
  select(year, patch, year_patch, control_treatment, exp_phase, bird_id, bird_species, mass, wing_length, notes) %>%
  filter(!is.na(wing_length)) %>%
  filter(!is.na(mass)) %>%
  filter(!grepl("comio antes de pesar|fed before weight|fed b/f", notes, ignore.case = TRUE)) 

#Check: after all this filtering, do I measurements pre and post? Interested in pre-to-post change, so need measurements in both periods
birds_missing_periods <- recaptures_experiment_measurements %>%
  group_by(bird_id) %>%
  summarise(num_periods = n()) %>%
  ungroup() %>%
  filter(num_periods < 2)

#Remove birds without pre and post measurements
recaptures_experiment_all_spp <- recaptures_experiment_measurements %>%
  filter(!(bird_id %in% birds_missing_periods$bird_id)) %>%
  mutate(bird_group = "all_spp")

#Create subset for GREH/VISA
recaptures_experiment_greh_visa <- recaptures_experiment_all_spp %>%
  filter(bird_species == "GREH" | bird_species == "VISA") %>%
  mutate(bird_group = "greh_visa")

#Combine subsets, add info about quantitative treatment
recaptures_experiment_for_analysis <- recaptures_experiment_all_spp %>%
  bind_rows(recaptures_experiment_greh_visa) %>%
  left_join(heto_removed) %>%
  mutate(log_calories_removed_per_ha = log(calories_removed_per_ha + 1)) %>%
  mutate(analysis = "body_mass",
         subanalysis = "body_mass") %>%
  select(analysis, subanalysis, bird_group, everything()) 

#How many individuals is this final dataset?
#All species
unique(recaptures_experiment_all_spp$bird_id) %>% length

#GREH/VISA
unique(recaptures_experiment_greh_visa$bird_id) %>% length
```

## Step 3: Using all capture data, plot log(mass) ~ log(wing length) for each species

Get capture data ready for allometric equation
```{r}
#Remove wing lengths for recaptures (id'ed in previous step), replace with average lengths for consistency
#Remove any birds without any wing measurements
#Remove any mass measurements where bird fed sugar water before mass was measured
#Add column indicating whether bird is included in recap dataset for Heliconia removal experiment (for plotting)
data_mass_wing_regressions <- captures_all_years %>%
  filter(!(bird_id %in% recaptures_experiment_measurements$bird_id)) %>% #Remove original data associated with recaptures (i.e., bird caught in pre and post)
  bind_rows(recaptures_experiment_measurements) %>%
  filter(!is.na(wing_length)) %>%
  filter(!is.na(mass)) %>%
  filter(!grepl("comio antes de pesar|fed before weight|fed b/f", notes, ignore.case = TRUE)) %>%
  mutate(experiment_bird = ifelse(bird_id %in% recaptures_experiment_for_analysis$bird_id, "yes", "no")) %>%
  mutate(log_mass = log(mass)) %>%
  mutate(log_wing_length = log(wing_length)) %>%
  select(year, patch, exp_phase, bird_species, bird_id, log_mass, log_wing_length, experiment_bird)
```

Write functions
```{r}
#Subset data to each species
subset_data <- function(dataset, bird_species){
  
  data <- dataset %>%
    filter(bird_species == {{ bird_species }})
  
  return(data)
  
}

#test_data = subset_data(dataset = data_mass_wing_regressions, bird_species = "VISA")

#Create model for mass ~ wing length
create_model <- function(dataset){
  
  model <- lm(log_mass ~ log_wing_length, data = dataset)
  return(model)
  
}

#Add residuals and predicted values
augment_data <- function(dataset, model_object){
  
  data <- dataset %>%
    select(-log_mass, -log_wing_length) %>%
    cbind(broom::augment(model_object)) 

  return(data)
  
}

#Find outliers using cook's distance
identify_outliers <- function(dataset, sample_size){
  
  data <- dataset %>%
    rename(cooksd = ".cooksd") %>%
    mutate(cooksd_cutoff = 8/sample_size) %>% #Usually the cutoff is 4/n, but that was kicking out too much data
    mutate(cooksd_outlier = ifelse(cooksd > cooksd_cutoff, 1, 0))

  return(data)
  
}

#Plot mass ~ wing length
#Goal: identify outliers
make_plot_outlier_id <- function(dataset, bird_species, group_id){
  
  plot <- ggplot(aes(y = log_mass, x = log_wing_length), data = dataset) +
    geom_point(aes(colour = factor(get(group_id))), size = 2) +
    geom_line(aes(y = .fitted, x = log_wing_length), data = dataset) + #This is exactly the same as using geom_smooth() with method LM
    labs(title = bird_species, x = "Log(wing length)", y = "Log(mass)") +
    scale_color_manual(values = c("gray80", "#BA0022")) +
    theme_bw(base_size = 18) +
    theme(legend.position = "none")

  return(plot)
  
}

#Plot mass ~ wing length
#Goal: plot allometric equations
make_plot <- function(dataset, bird_species, intercept, slope, group_id){
  
  intercept = round(intercept, digits = 2)
  slope = round(slope, digits = 2)
  label = bquote("y ="~.(intercept)~"+"~.(slope)~"x") #Without parse(), it is a call.
  label_df <- data.frame(label = deparse(label))
  
  plot <- ggplot(aes(y = log_mass, x = log_wing_length), data = dataset) +
    geom_point(aes(colour = factor(get(group_id))), size = 2) +
    geom_line(aes(y = .fitted, x = log_wing_length), data = dataset) + #This is exactly the same as using geom_smooth() with method LM
    geom_text(aes(x = Inf, y = Inf, label = label), data = label_df, parse = TRUE, hjust = 1.25, vjust = 1.5, size = 6, colour = "black") + #Upper right
    labs(title = bird_species, x = "Log(wing length)", y = "Log(mass)") +
    scale_color_manual(values = c("gray80", "#BA0022")) +
    theme_bw(base_size = 18) +
    theme(legend.position = "none")

  return(plot)
  
}
```

Apply functions
```{r}
#1. Identify hummingbird species that I will need regressions for (the species I have recaptures for)
#2. Subset larger capture dataset to species of interest
#3. Only run regressions for species that have >30 observations
#4. Run regression
#5. Add residuals and predicted values from model, using broom::augment
mass_wing_regressions <- recaptures_experiment_for_analysis %>%
  distinct(bird_species) %>%
  left_join(bird_species_names) %>%
  mutate(data = map(bird_species, ~subset_data(dataset = data_mass_wing_regressions, bird_species = .))) %>% 
  mutate(num_rows = map(data, function(.){unlist(dim(.)[[1]])})) %>%
  filter(num_rows > 30) %>% #Remove species that don't have > 30 observations for regression
  mutate(mass_wing_model = map(data, create_model)) %>%
  mutate(data_augment = pmap(list(data, mass_wing_model), ~augment_data(dataset = ..1, model_object = ..2)))
```

Identify and remove outliers from regression
```{r}
#Identify outliers in mass ~ wing length regressions, with the goal of removing sketchy points (in lieu of going back to paper datasheets)
outliers <- mass_wing_regressions %>%
  mutate(data_augment_outliers = pmap(list(data_augment, num_rows), ~identify_outliers(dataset = ..1, sample_size = ..2))) %>%
  mutate(plot_outlier = pmap(list(data_augment_outliers, bird_common_name), ~make_plot_outlier_id(dataset = ..1, bird_species = ..2, group_id = "cooksd_outlier")))

#Sample plot of what outliers look like
outliers$plot_outlier[[1]]

#As data frame...
outliers_df <- bind_rows(outliers$data_augment_outliers) %>%
  filter(cooksd_outlier == 1)

#Are there any birds ID'ed as outliers that are in my recap dataset? NO.
check01 <- outliers_df %>%
  filter(bird_id %in% recaptures_experiment_for_analysis$bird_id)

#Remove outliers
data_mass_wing_regressions_without_outliers <- data_mass_wing_regressions %>%
  filter(!(bird_id %in% outliers_df$bird_id))
```

Re-run regressions using data without outliers
```{r, fig.height = 11.5, fig.width = 15}
mass_wing_regressions_without_outliers <- recaptures_experiment_for_analysis %>%
  distinct(bird_species) %>%
  left_join(bird_species_names) %>%
  mutate(data = map(bird_species, ~subset_data(dataset = data_mass_wing_regressions_without_outliers, bird_species = .))) %>% 
  mutate(num_rows = map(data, function(.){unlist(dim(.)[[1]])})) %>%
  filter(num_rows > 30) %>% #Remove species that don't have > 30 observations for regression
  mutate(mass_wing_model = map(data, create_model)) %>%
  mutate(mass_wing_intercept = map(mass_wing_model, function(.){coefficients(.)[[1]]})) %>%
  mutate(mass_wing_slope = map(mass_wing_model, function(.){coefficients(.)[[2]]})) %>%
  mutate(data_augment = pmap(list(data, mass_wing_model), ~augment_data(dataset = ..1, model_object = ..2))) %>%
  mutate(plot_recaps = pmap(list(data_augment, bird_common_name, mass_wing_intercept, mass_wing_slope), ~make_plot(dataset = ..1, bird_species = ..2, intercept = ..3, slope = ..4, group_id = "experiment_bird"))) %>%
  arrange(desc(as.numeric(num_rows))) #Re-order so that plots are ordered by amount of data

mass_wing_regressions_without_outliers$plot_recaps[[1]]
mass_wing_regressions_without_outliers$mass_wing_model[[1]] %>% summary()
mass_wing_regressions_without_outliers$num_rows[[1]]

#Combine plots of mass-wing regressions, for Supp Material
mass_wing_regressions_multiplot <- ggpubr::ggarrange(plotlist = mass_wing_regressions_without_outliers$plot_recaps)
mass_wing_regressions_multiplot
```

## Step 4: Calculate relative body mass
```{r}
#Get relative mass for birds in experiment (2016-2018)
relative_mass_for_analysis <- bind_rows(mass_wing_regressions_without_outliers$data_augment) %>%
  filter(experiment_bird == "yes") %>%
  rename(predicted_log_mass = ".fitted") %>%
  mutate(relative_mass = log_mass/predicted_log_mass) %>%
  select(-starts_with(".")) %>%
  select(-experiment_bird)

#Add relative mass to rest of data
data_for_analysis <- recaptures_experiment_for_analysis %>%
  left_join(relative_mass_for_analysis)
```

## Step 5: Analyze how relative body mass changes (or not) in response to *Heliconia* removal

**Prepare for analysis**
```{r}
#Create tables that will be filled in with models and model results
analysis = c("body_mass")
subanalysis = c("body_mass")
bird_group = c("all_spp", "greh_visa")
model_type = c("categorical", "quantitative")
model_name = c("control/treatment * pre/post", "calories removed/ha * pre/post")
treatment_variable = c("control_treatment", "log_calories_removed_per_ha")
model_types_names = data.frame(model_type = model_type, model_name = model_name, treatment_variable = treatment_variable)

base_table <- crossing(analysis, subanalysis, bird_group, model_type) %>% 
  left_join(model_types_names) %>%
  mutate(yvar = "relative_mass") %>%
  arrange(analysis, subanalysis, model_type, bird_group) %>%
  mutate(model_number = 1:length(yvar)) %>%
  mutate(title = paste(analysis, subanalysis, bird_group, sep = "_")) %>%
  mutate(subtitle = yvar) %>%
  select(model_number, analysis, yvar, subanalysis, everything())
```

**Filter master dataset to relevant sub-dataset & scale continuous predictors**
```{r}
subset_data <- function(dataset, analysis, subanalysis, bird_group){
  
  dataset <- dataset %>%
    filter(analysis == {{ analysis }}) %>%
    filter(subanalysis == {{ subanalysis}}) %>%
    filter(bird_group == {{ bird_group }}) %>%
    mutate(across(c(log_calories_removed_per_ha), ~c(scale(.)))) %>%
    mutate(exp_phase = factor(exp_phase, levels = c("pre", "post")))
      
  return(dataset)
  
}

#Create datasets (will add models in later step)
data <- base_table %>%
    mutate(data = pmap(list(analysis, subanalysis, bird_group), ~subset_data(dataset = data_for_analysis, analysis = ..1, subanalysis = ..2, bird_group = ..3)))
```

**Initial model fitting**
```{r, include = FALSE}
fit_data01 <- data_for_analysis %>%
  filter(analysis == "body_mass" & subanalysis == "body_mass" & bird_group == "all_spp")

#Can I include full random effect structure?
m1 <- glmmTMB(relative_mass ~ control_treatment * exp_phase + (1|patch/year_patch/bird_id), data = fit_data01, family = "gaussian", control = glmmTMBControl(optimizer=optim, optArgs=list(method="BFGS")))
m2 <- glmmTMB(relative_mass ~ control_treatment * exp_phase + (1|year_patch/bird_id), data = fit_data01, family = "gaussian", control = glmmTMBControl(optimizer=optim, optArgs=list(method="BFGS")))
m3 <- glmmTMB(relative_mass ~ control_treatment * exp_phase + (1|bird_id), data = fit_data01, family = "gaussian", control = glmmTMBControl(optimizer=optim, optArgs=list(method="BFGS")))

#No; run into convergence problems if try to include RE beyond bird_id

#Do assumptions for this model look OK?
performance::check_model(m3)

#Yes; look OK, but will check for all models later on.
```

**Fit models**
```{r}
create_model <- function(dataset, analysis, subanalysis, treatment_variable){

  model <- glmmTMB(as.formula(paste0("relative_mass ~ ", treatment_variable, " * exp_phase + (1|bird_id)")), data = dataset, na.action = na.omit, family = "gaussian")
      
  return(model)
  
}

#Create models
models <- data %>%
   mutate(model_quietly = pmap(list(data, analysis, subanalysis, treatment_variable), quietly(function(a, b, c, d){model <- create_model(dataset = a, analysis = b, subanalysis = c, treatment_variable = d)})),
          model = map(model_quietly, get_result),
          warning = map(model_quietly, get_warning), 
          convergence = map(model, check_convergence)) %>%
  mutate(xvar_table = map(model, ~data.frame(xvar = all.vars(terms(.))[-1])))

#Make sure there weren't any problems + check to make sure table of predictor variables is correct
models$warning
models$convergence
models$xvar_table
```

**Check assumptions**
```{r, fig.show = FALSE}
assumption_checks <- models %>%
    mutate(model_check = map(model, performance::check_model))

assumption_checks$model_check
```

**Get results together for export**
```{r}
results <- models %>%
  mutate(summary = map(model, summary),
         summary_tidy = map(model, broom.mixed::tidy),
         confint_wald = map(model, ~calculate_ci(., method_name = "wald"))) %>%
  mutate(num_obs = map(model, get_number_obs),
         num_levels = map(model, get_number_re_levels),
         num_reps = map(data, ~get_sample_size(data = ., vars = c("year", "patch", "control_treatment"), grouping_var = "control_treatment")))
```

## Step 6: Export results
```{r}
#Data that I analyzed here
write.csv(data_for_analysis, "../../data/export/for_analysis/Relative_body_mass.csv", row.names = FALSE)

#Multiplot of log(mass) ~ log(wing length) regressions
path_name <- c("../../results/statistical_analysis/figures")
ggsave(filename = file.path(path_name, "FigS2_Allometric_regression_plots.png"), plot = mass_wing_regressions_multiplot, height = 11.5, width = 15)

#Responses to Heliconia removal
saveRDS(results, "../../results/statistical_analysis/rds_files/Body_mass_results.rds")
```
